+++ paper/src/main/java/com/github/maxopoly/finale/combat/knockback/StandardKnockback.java
@@ -15,7 +15,7 @@ import org.bukkit.util.Vector;
 public class StandardKnockback implements KnockbackStrategy {
 
 	@Override
-	public void handleKnockback(Player attacker, Entity entity, int knockbackLevel) {
+	public void handleKnockback(Player attacker, Entity entity, float knockbackLevel) {
 		CombatConfig config = Finale.getPlugin().getManager().getCombatConfig();
 		SprintHandler sprintHandler = Finale.getPlugin().getManager().getSprintHandler();
